refractor common txns
