+++ executor/src/lib.rs
@@ -5,9 +5,9 @@ pub use block_executor::block_execute;
 pub use executor::*;
 pub use transaction_builder::{
     build_accept_coin_txn, build_mint_txn, build_transfer_txn, build_transfer_txn_by_coin_type,
-    encode_create_account_script, encode_transfer_script, peer_to_peer_txn_sent_as_association,
+    create_signed_txn_with_association_account, encode_create_account_script,
+    encode_transfer_script, peer_to_peer_txn_sent_as_association, TXN_RESERVED,
 };
-pub use vm_runtime::common_transactions::create_signed_txn_with_association_account;
 
 mod block_executor;
 mod executor;
