+++ executor/src/transaction_builder.rs
@@ -1,14 +1,18 @@
+use starcoin_types::account_address::AccountAddress;
 use starcoin_types::language_storage::TypeTag;
-use starcoin_types::{
-    account_address::AccountAddress,
-    transaction::{RawUserTransaction, SignedUserTransaction, Transaction},
+use starcoin_types::transaction::{
+    RawUserTransaction, Script, SignedUserTransaction, Transaction, TransactionArgument,
+    TransactionPayload,
 };
+use starcoin_vm_types::account_config;
 use starcoin_vm_types::account_config::stc_type_tag;
-use starcoin_vm_types::transaction::{Script, TransactionPayload};
-use starcoin_vm_types::transaction_argument::TransactionArgument;
 use std::time::Duration;
-use vm_runtime::common_transactions::DEFAULT_EXPIRATION_TIME;
+use vm_runtime::genesis::GENESIS_KEYPAIR;
 use vm_runtime::transaction_scripts::{ACCEPT_COIN_TXN, CREATE_ACCOUNT_TXN, PEER_TO_PEER_TXN};
+
+//TODO move to transaction_builder crate.
+pub const DEFAULT_EXPIRATION_TIME: u64 = 40_000;
+pub const TXN_RESERVED: u64 = 2_000_000;
 
 pub fn build_mint_txn(
     addr: AccountAddress,
@@ -171,7 +175,29 @@ pub fn peer_to_peer_txn_sent_as_association(
         vec![stc_type_tag()],
         args,
         seq_num,
-        vm_runtime::common_transactions::TXN_RESERVED,
+        TXN_RESERVED,
         1,
     )
+}
+
+pub fn create_signed_txn_with_association_account(
+    program: Vec<u8>,
+    ty_args: Vec<TypeTag>,
+    args: Vec<TransactionArgument>,
+    sequence_number: u64,
+    max_gas_amount: u64,
+    gas_unit_price: u64,
+) -> SignedUserTransaction {
+    RawUserTransaction::new(
+        account_config::association_address(),
+        sequence_number,
+        TransactionPayload::Script(Script::new(program, ty_args, args)),
+        max_gas_amount,
+        gas_unit_price,
+        // TTL is 86400s. Initial time was set to 0.
+        Duration::from_secs(DEFAULT_EXPIRATION_TIME),
+    )
+    .sign(&GENESIS_KEYPAIR.0, GENESIS_KEYPAIR.1.clone())
+    .unwrap()
+    .into_inner()
 }
