+++ testsuite/tests/steps/transaction.rs
@@ -4,6 +4,7 @@
 use crate::MyWorld;
 use anyhow::Error;
 use cucumber::{Steps, StepsBuilder};
+use starcoin_executor::TXN_RESERVED;
 use starcoin_logger::prelude::*;
 use starcoin_rpc_client::{RemoteStateReader, RpcClient};
 use starcoin_state_api::AccountStateReader;
@@ -12,7 +13,6 @@ use starcoin_types::account_config;
 use starcoin_types::transaction::authenticator::AuthenticationKey;
 use starcoin_types::transaction::helpers::TransactionSigner;
 use starcoin_types::transaction::{RawUserTransaction, SignedUserTransaction};
-use starcoin_vm_runtime::common_transactions::TXN_RESERVED;
 use starcoin_wallet_api::WalletAccount;
 use std::time::Duration;
