+++ vm/functional-tests/Cargo.toml
@@ -27,6 +27,8 @@ starcoin-logger = { path = "../../commons/logger"}
 starcoin-statedb = { path = "../../state/statedb" }
 scs = { package="starcoin-canonical-serialization", path = "../../commons/scs"}
 starcoin-move-compiler = { path = "../../vm/compiler"}
+executor = { package="starcoin-executor", path = "../../executor"}
+
 
 [[test]]
 name = "testsuite"
